+++ drivers/mssql/adapters.js
@@ -1,4 +1,5 @@
-const mssql = require('mssql');
+const ConnectionPool = require('tedious-connection-pool');
+const Connection = require('tedious').Connection;
 class Adapter {
     constructor(nqb) {
         // Verify that an instance of Node QueryBuilder was passed in
@@ -16,45 +17,91 @@ class Adapter {
 
         // Enable debugging if necessary
         this.debugging = false;
-        if (this.nqb.settings.hasOwnProperty('qb_debug') && this.nqb.settings.qb_debug === true) {
+        if (this.nqb.settings.hasOwnProperty('debug') && this.nqb.settings.debug === true) {
             this.debugging = true;
-            delete this.nqb.settings.qb_debug;
+            delete this.nqb.settings.debug;
         }
 
         // Verify that required fields are provided...
         if (Object.keys(this.nqb.settings).length === 0) throw new Error("No connection information provided!");
         if (!this.nqb.settings.hasOwnProperty('host')) this.nqb.settings.host = 'localhost';
-        if (!this.nqb.settings.hasOwnProperty('user')) throw new Error("No user property provided. Hint: It can be NULL");
-        //if (!nqb.settings.hasOwnProperty('password')) throw new Error("No connection password provided. Hint: It can be NULL");
+        if (!this.nqb.settings.hasOwnProperty('user')) { console.log("Settings:", this.nqb.settings); throw new Error("No user property provided. Hint: It can be NULL"); }
 
         this.map_connection_settings();
+
     }
 
     // ****************************************************************************
     // Map generic NQB connection settings to mssql's format
     // ****************************************************************************
     map_connection_settings() {
-        this.connection_settings = {
-            server: this.nqb.settings.host,
-            user: this.nqb.settings.user,
-            password: this.nqb.settings.password
+        const settings = Object.assign({}, this.nqb.settings);
+
+        this._connection_settings = {
+            server: settings.host,
+            userName: settings.user,
+            password: settings.password,
+            options: {
+                port: 1433,
+                encrypt: false,
+                rowCollectionOnRequestCompletion: true,
+                fallbackToDefaultDb: false,
+                debug: {
+                    packet: this.debugging,
+                    data: this.debugging,
+                    payload: this.debugging,
+                    token: this.debugging,
+                }
+            }
+        };
+
+        if (settings.hasOwnProperty('database')) {
+            this._connection_settings.options.database = settings.database;
+            delete settings.database;
         }
-        if (this.nqb.settings.hasOwnProperty('database')) {
-            this.connection_settings.database = this.nqb.settings.database;
-            delete this.nqb.settings.database
-        }
-        if (this.nqb.settings.hasOwnProperty('port')) {
-            this.connection_settings.port = this.nqb.settings.port;
-            delete this.nqb.settings.port
+        if (settings.hasOwnProperty('port')) {
+            this._connection_settings.options.port = settings.port;
+            delete settings.port;
         }
 
-        // Remove mapped settings:
-        delete this.nqb.settings.host
-        delete this.nqb.settings.user
-        delete this.nqb.settings.password
+        // Remove mapped connection settings:
+        delete settings.host;
+        delete settings.user;
+        delete settings.password;
 
-        // Merge any driver-specific settings into connection settings
-        this.connection_settings = Object.assign(this.connection_settings, this.nqb.settings);
+        // Set default pool settings
+        this.pool_settings = {
+            min: 10,
+            max: 10,
+            acquireTimeout: 60000,
+            log: this.debugging,
+        };
+
+        // Override default pool settings
+        if (settings.hasOwnProperty('pool_size')) {
+            this.pool_settings.max = settings.pool_size;
+            delete settings.pool_size;
+        }
+        if (settings.hasOwnProperty('pool_min')) {
+            this.pool_settings.min = settings.pool_min;
+            delete settings.pool_min;
+        }
+        if (settings.hasOwnProperty('acquireTimeout')) {
+            this.pool_settings.acquireTimeout = settings.acquireTimeout;
+            delete settings.acquireTimeout;
+        }
+
+
+        if (settings.hasOwnProperty('options') && typeof settings.options === 'object') {
+            let options = this._connection_settings.options;
+            options = Object.assign(options, settings.options);
+            options.debug = this._connection_settings.options.debug;
+            this._connection_settings.options = options;
+            delete settings.options;
+        }
+
+        // Merge any additional driver-specific settings into connection settings
+        this._connection_settings = Object.assign(this._connection_settings, settings);
     }
 
     // ****************************************************************************
@@ -67,7 +114,7 @@ class Adapter {
         try {
             return require('./query_builder.js').QueryBuilder();
         } catch(e) {
-            throw new Error("Couldn't load the QueryBuilder library for " + this.nqb.driver + ": " + e);
+            throw new Error(`Couldn't load the QueryBuilder library for ${this.nqb.driver}: ${e}`);
         }
     }
 
@@ -83,7 +130,7 @@ class Adapter {
         try {
             return require('./query_exec.js').QueryExec(qb, conn);
         } catch(e) {
-            throw new Error("Couldn't load the QueryExec library for " + this.nqb.driver + ": " + e);
+            throw new Error(`Couldn't load the QueryExec library for ${this.nqb.driver}: ${e}`);
         }
     }
 }
@@ -109,23 +156,30 @@ class Single extends Adapter {
         }
         // Otherwise, let's create a new connection
         else {
-            this._connection = mssql.connect(this.connection_settings);
+            const self = this;
+            function SQLConnection() {};
+            SQLConnection.prototype.connect = function(callback) {
+                this.connection = new Connection(self._connection_settings);
+                this.connection.on('error', callback);
+                this.connection.on('connect', callback);
+                return this.connection;
+            }
+            this.sql_connection = new SQLConnection();
         }
 
         this.qb = this.get_query_builder();
-        this.qe = this.get_query_exec(this.qb, this._connection);
+        this.qe = this.get_query_exec(this.qb, this.sql_connection);
 
         const self = this;
 
         return Object.assign({
             connection_settings: function() {
-                return self.connection_settings;
+                return {connection_settings: self._connection_settings, pool_settings: self.pool_settings};
             },
 
             connect: function(callback) {
-                return self._connection.then(err => {
-                    return callback(err);
-                });
+                if (self._connection) return self._connection;
+                self._connection = self.sql_connection.connect(callback);
             },
 
             connection: function() {
@@ -141,12 +195,17 @@ class Single extends Adapter {
             },
 
             disconnect: function(callback) {
-                return self.connection.end(callback);
+                if (self.pool) {
+                    self.pool.drain();
+                } else {
+                    self._connection.close();
+                }
+                if (callback && typeof callback === 'function') callback(null);
             },
 
             release: function() {
                 if (!self.pool) throw new Error("You cannot release a non-pooled connection from a connection pool!");
-                self.pool.releaseConnection(self.connection);
+                self.pool.release(self._connection);
             }
         }, self.qb, self.qe);
     }
@@ -164,14 +223,11 @@ class Pool extends Adapter {
         // Create pool for node-querybuilder object if it doesn't already have one.
         if (!this.nqb.hasOwnProperty('pool') || this.nqb.pool.length === 0) {
             // Create connection Pool
-            this.nqb.pool = new mssql.ConnectionPool(this.connection_settings, err => {
-                if (this.debugging === true) {
-                    if (err) {
-                        console.error(err);
-                    } else {
-                        console.log('mssql connection pool created');
-                    }
-                }
+            const ps = Object.assign({}, this.pool_settings);
+            const cs = Object.assign({}, this._connection_settings);
+            this.nqb.pool = new ConnectionPool(ps, cs);
+            this.nqb.pool.on('error', err => {
+                if (this.debugging === true) console.error(err);
             });
         }
 
@@ -188,20 +244,21 @@ class Pool extends Adapter {
                     throw new Error(error_msg);
                 }
 
-                self.nqb.pool.getConnection((err, connection) => {
+                self.nqb.pool.acquire((err, connection) => {
                     if (err) throw err;
-                    const adapter = new Single(nqb, {
+                    const adapter = new Single(self.nqb, {
                         pool: {
                             pool: self.nqb.pool,
-                            connection: connection
+                            connection: connection,
                         }
                     });
 
                     callback(adapter);
                 });
+
             },
             disconnect: function(callback) {
-                self.nqb.pool.close(callback);
+                self.nqb.pool.drain(callback);
             }
         }
     }